+++ test/runtests.jl
@@ -20,4 +20,5 @@ import ArraysOfArrays, SpecialFunctions
     include("structured.jl")
     include("multivariate.jl")
     include("adapters.jl")
+    include("chain_adapters.jl")
 end